script can take in custom input
Added command-line arguments so the script can take in our custom .tsv input. This input is the output .tsv file and will be generated by running the scraper code and then the clean code to filter the tsv file. The .tsv file will be read into a list where each element in the list will undergo a sentiment analysis.